requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
crontab==1.0.1
diff-match-patch==20230430
Django==5.0.7
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-import-export==4.1.1
django-redis==5.4.0
django-rq==2.10.2
django-timezone-field==7.0
djangorestframework==3.15.2
flower==2.0.1
freezegun==1.5.1
humanize==4.10.0
kombu==5.3.7
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.7
rq==1.16.2
rq-scheduler==0.13.1
rq-win @ git+https://github.com/michaelbrooks/rq-win.git@c6e65e38f8a6ca99a0c8e0dd6fe35f9b92837cca
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
times==0.7
tornado==6.4.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
